Making rejet and accept buttons components
