+++ package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^5.0.0-beta3",
     "bootstrap-icons": "^1.5.0",
     "jquery": "^3.6.0",
+    "moment": "^2.29.1",
     "node-sass": "^6.0.1",
     "popper.js": "^1.16.1",
     "react": "^17.0.2",
